Module bblog::app

Structs

AllPostsProps

Props for the AllPosts component.

AllPostsPropsBuilder

Builder for AllPostsProps instances.

AppProps

Props for the App component.

AppPropsBuilder

Builder for AppProps instances.

AuthorLinkProps

Props for the AuthorLink component.

AuthorLinkPropsBuilder

Builder for AuthorLinkProps instances.

DeletePost

Serialized arguments for the delete_post server function.

GetAllPosts

Serialized arguments for the get_all_posts server function.

GetAllUserSeries

Serialized arguments for the get_all_user_series server function.

GetPost

Serialized arguments for the get_post server function.

GetSeries

Serialized arguments for the get_series server function.

GetSubscriptionsPosts

Serialized arguments for the get_subscriptions_posts server function.

GetUserFromID

Serialized arguments for the get_user_from_id server function.

GetUserFromUsername

Serialized arguments for the get_user_from_username server function.

GetUserPosts

Serialized arguments for the get_user_posts server function.

GetUserSeriesFromName

Serialized arguments for the get_user_series_from_name server function.

LoginProps

Props for the [Login] component.

LoginPropsBuilder

Builder for LoginProps instances.

LogoutProps

Props for the [Logout] component.

LogoutPropsBuilder

Builder for LogoutProps instances.

NewPost

Serialized arguments for the new_post server function.

NewPostDraft

Serialized arguments for the new_post_draft server function.

NewPostDraftProps

Props for the [NewPostDraft] component.

NewPostDraftPropsBuilder

Builder for NewPostDraftProps instances.

NewPostProps

Props for the [NewPost] component.

NewPostPropsBuilder

Builder for NewPostProps instances.

NewSeries

Serialized arguments for the new_series server function.

NewSeriesProps

Props for the [NewSeries] component.

NewSeriesPropsBuilder

Builder for NewSeriesProps instances.

Post

PostCardProps

Props for the PostCard component.

PostCardPropsBuilder

Builder for PostCardProps instances.

PostListProps

Props for the PostList component.

PostListPropsBuilder

Builder for PostListProps instances.

PostProps

Props for the [Post] component.

PostPropsBuilder

Builder for PostProps instances.

Series

SeriesLinkProps

Props for the SeriesLink component.

SeriesLinkPropsBuilder

Builder for SeriesLinkProps instances.

SignupProps

Props for the [Signup] component.

SignupPropsBuilder

Builder for SignupProps instances.

User

UserFirstAndLastNameProps

Props for the UserFirstAndLastName component.

UserFirstAndLastNamePropsBuilder

Builder for UserFirstAndLastNameProps instances.

UserPostsProps

Props for the UserPosts component.

UserPostsPropsBuilder

Builder for UserPostsProps instances.

UserProfileLinkProps

Props for the UserProfileLink component.

UserProfileLinkPropsBuilder

Builder for UserProfileLinkProps instances.

UserProfileProps

Props for the UserProfile component.

UserProfilePropsBuilder

Builder for UserProfileProps instances.

UserSeriesDropDownProps

Props for the UserSeriesDropDown component.

UserSeriesDropDownPropsBuilder

Builder for UserSeriesDropDownProps instances.

Functions

AllPosts

PostList containing every post in the database

App

Main app component for rendering all routes

AuthorLink

Displays a user’s profile picture alongside a link to their profile, most commonly used for displaying authors in PostCard components

Login

Displays a login page for logging a user into BBlog

Logout

Displays a logout page for logging a user out of BBlog

NewPost

Provides the user with input fields to make a new Post of their own

NewPostDraft

Provides the user with input fields to make a draft of a Post

NewSeries

Provides the user with input fields to make a new Series of their own

Post

Displays a Post based on the post_id in the URL route

PostCard

Displays info about a Post alongside a thumbnail to entice a user to read it

PostList

Generic PostList component for rendering a vector of posts as PostCard components

SeriesLink

Displays a user’s series based its id

Signup

Displays a login page for creating an account on BBlog

UserFirstAndLastName

Displays a user’s first and last name based on their id

UserPosts

PostList containing every post by a specified user based on their id in the database

UserProfile

Displays a user’s information alongside their most recent posts

UserProfileLink

Displays a user’s first and last name as a link to their profile page

UserSeriesDropDown

Displays a dropdown picker of user’s series alongside an input and button for a specified User by id

delete_post

Deletes a post based on its id from the posts table in the database

get_all_posts

Gets all posts entries from the posts table in the database

get_all_user_series

Gets all series entries for a specific user based on their id from the series table in the database

get_post

Gets a post entry from the posts table in the database

get_series

Gets a series entry from the series table in the database

get_subscriptions_posts

Gets all post entries from a specific user’s subscriptions based on their id from the posts table in the database

get_user_from_id

Gets a specific user based on their id from the users table in the database

get_user_from_username

Gets a specific user based on their username from the users table in the database

get_user_posts

Gets all post entries for a specific user based on their id from the posts table in the database

get_user_series_from_name

Gets a series for a specific user based on their username from the series table in the database

new_post

Inserts a new post into the posts table in the database with its posted flag set to false, to save it as a draft that only the current user can see

new_post_draft

Inserts a new post into the posts table in the database with its posted flag set to false, to save it as a draft that only the current user can see

new_series

Inserts a new series into the series table in the database